@@ -1,7 +1,8 @@
+import { Request, Response, NextFunction } from 'express'
+import mockedUser from '../tests/mocks/user'
 const isTestMode = process.env.NODE_ENV === 'test'
-const mockedUser = require('../test/mocks').user
 
-module.exports = (req, res, next) => {
+export default (req: Request, res: Response, next: NextFunction) => {
   // make sure we can still access routes in tests, create mock user
   if (isTestMode) {
     req.user = mockedUser
